+++ crates/deskulpt-core/src/commands/emit_on_render_ready.rs
@@ -1,13 +1,13 @@
 use tauri::{command, AppHandle, Runtime};
 
 use super::error::CmdResult;
-use crate::states::StatesExtRenderReady;
+use crate::states::StatesExtInitialRender;
 
-/// Emit a render event to the canvas when it is ready.
+/// Emit a `render-widgets` event to the canvas when it is ready.
 ///
 /// ### Errors
 ///
-/// - Failed to emit the render event to the canvas.
+/// - Failed to emit the `render-widgets` event to the canvas.
 #[command]
 pub async fn emit_on_render_ready<R: Runtime>(
     app_handle: AppHandle<R>,
