+++ crates/deskulpt-core/src/commands/set_render_ready.rs
@@ -1,13 +1,13 @@
 use tauri::{command, AppHandle, Runtime};
 
 use super::error::CmdResult;
-use crate::states::StatesExtRenderReady;
+use crate::states::StatesExtInitialRender;
 
 /// Set the canvas as ready to render widgets.
 ///
 /// ### Errors
 ///
-/// - Failed to emit the render event to the canvas.
+/// - Failed to emit the `render-widgets` event to the canvas.
 #[command]
 pub async fn set_render_ready<R: Runtime>(app_handle: AppHandle<R>) -> CmdResult<()> {
     Ok(app_handle.set_render_ready()?)
