+++ crates/deskulpt-core/src/states/mod.rs
@@ -1,12 +1,12 @@
 //! Deskulpt runtime state management.
 
 mod canvas_click_through;
-mod render_ready;
+mod initial_render;
 mod widget_config_map;
 
 #[doc(hidden)]
 pub use canvas_click_through::StatesExtCanvasClickThrough;
 #[doc(hidden)]
-pub use render_ready::StatesExtRenderReady;
+pub use initial_render::StatesExtInitialRender;
 #[doc(hidden)]
 pub use widget_config_map::StatesExtWidgetConfigMap;
